refactor(backend): Entity ID 타입을 int에서 Long으로 변경
  - Account, Profile, RefreshToken 엔티티 ID 타입 변경 (int → Long)
  - Repository 제네릭 타입 변경 (Integer → Long)
  - Service 메서드 파라미터 타입 업데이트
  - DTO 필드 타입 변경 (int → Long)
  - Lombok 접근 제어자 추가 (PROTECTED, PRIVATE)
  - 불필요한 주석 제거 (// fields, // methods)
  - 테스트 코드 전면 업데이트 (anyInt() → anyLong(), 1 → 1L)

+++ backend/src/main/java/com/project/blog/domain/account/dto/res/UserInfoResDto.java
@@ -12,7 +12,7 @@ import lombok.Getter;
 @Getter
 @Builder
 public class UserInfoResDto {
-        private int id;
+        private Long id;
         private String username;
         private String email;
         private RoleType roleType;
@@ -22,7 +22,7 @@ public class UserInfoResDto {
         @Getter
         @Builder
         public static class ProfileInfo {
-                private int id;
+                private Long id;
                 private String nickname;
                 private String blogName;
                 private String avatar;
