+++ backend/src/main/java/com/project/blog/domain/profile/repository/ProfileRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.project.blog.domain.profile.entity.Profile;
 
-public interface ProfileRepository extends JpaRepository<Profile, Integer> {
+public interface ProfileRepository extends JpaRepository<Profile, Long> {
     boolean existsByNickname(String nickname);
 
     boolean existsByBlogName(String blogName);
